refactor: 승률 계산 및 랭킹 산정 리팩토링
- 리팩토링 진행
